+++ SBBDesignSystemMobileSwiftUIDemo/Views/DetailViews/NotificationDemo.swift
@@ -12,8 +12,10 @@ struct NotificationDemo: View {
     @State var isPresented2: Bool = true
     @State var isPresented3: Bool = true
     @State var showMoreInfo: Bool = false
+    @State var retry: Bool = false
     @State var selectedStatus: StatusType = .alert
     
+    @State var canBeClosed: Bool = false
     @State var enableCloseAfterSeconds: Bool = false
     @State var closeAfterSeconds: Int = 3
     
@@ -38,20 +40,27 @@ struct NotificationDemo: View {
                     if enableMaxNumberLines {
                         SBBUpDnCounterView(label: Text("Max number of lines"), value: $maxNumberLines)
                     }
-                    SBBCheckBox(isOn: $enableCloseAfterSeconds, text: Text("Close the notification after a while"), showBottomLine: enableCloseAfterSeconds)
-                    if enableCloseAfterSeconds {
-                        SBBUpDnCounterView(label: Text("Close after x seconds"), value: $closeAfterSeconds, showBottomLine: false)
+                    
+                    SBBCheckBox(isOn: $canBeClosed, text: Text("Notification can be closed"), showBottomLine: canBeClosed)
+                    
+                    if canBeClosed {
+                        SBBCheckBox(isOn: $enableCloseAfterSeconds, text: Text("Close the notification after a while"), showBottomLine: enableCloseAfterSeconds)
+                        if enableCloseAfterSeconds {
+                            SBBUpDnCounterView(label: Text("Close after x seconds"), value: $closeAfterSeconds, showBottomLine: false)
+                        }
                     }
                 }
-            
+                
                 Text("With title and text")
-                if enableCloseAfterSeconds {
+                if !canBeClosed {
+                    SBBNotification(statusType: selectedStatus, title: Text("Title"), text: Text("Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore."), maxNumberLines: enableMaxNumberLines ? maxNumberLines : nil)
+                } else if enableCloseAfterSeconds {
                     SBBNotification(isPresented: $isPresented1, statusType: selectedStatus, title: Text("Title"), text: Text("Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore."), maxNumberLines: enableMaxNumberLines ? maxNumberLines : nil, closeAfterSeconds: closeAfterSeconds)
                 } else {
-                    SBBNotification(isPresented: $isPresented1, statusType: selectedStatus, title: Text("Title"), text: Text("Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore."), maxNumberLines: enableMaxNumberLines ? maxNumberLines : nil, closeAfterSeconds: nil)
+                    SBBNotification(isPresented: $isPresented1, statusType: selectedStatus, title: Text("Title"), text: Text("Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore."), maxNumberLines: enableMaxNumberLines ? maxNumberLines : nil)
                 }
                 
-                if !isPresented1 {
+                if !isPresented1 && canBeClosed {
                     HStack {
                         Spacer()
                         Button(action: {
@@ -65,12 +74,14 @@ struct NotificationDemo: View {
                 }
                 
                 Text("Without title")
-                if enableCloseAfterSeconds {
+                if !canBeClosed {
+                    SBBNotification(statusType: selectedStatus, text: Text("Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore."), maxNumberLines: enableMaxNumberLines ? maxNumberLines : nil)
+                } else if enableCloseAfterSeconds {
                     SBBNotification(isPresented: $isPresented2, statusType: selectedStatus, text: Text("Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore."), maxNumberLines: enableMaxNumberLines ? maxNumberLines : nil, closeAfterSeconds: closeAfterSeconds)
                 } else {
-                    SBBNotification(isPresented: $isPresented2, statusType: selectedStatus, text: Text("Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore."), maxNumberLines: enableMaxNumberLines ? maxNumberLines : nil, closeAfterSeconds: nil)
+                    SBBNotification(isPresented: $isPresented2, statusType: selectedStatus, text: Text("Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore."), maxNumberLines: enableMaxNumberLines ? maxNumberLines : nil)
                 }
-                if !isPresented2 {
+                if !isPresented2 && canBeClosed {
                     HStack {
                         Spacer()
                         Button(action: {
@@ -84,16 +95,20 @@ struct NotificationDemo: View {
                 }
                 
                 Text("With more info")
-                if enableCloseAfterSeconds {
+                if !canBeClosed {
+                    SBBNotification(statusType: selectedStatus, title: Text("Title"), text: Text("Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore."), maxNumberLines: enableMaxNumberLines ? maxNumberLines : nil, onMoreInfo: {
+                        self.showMoreInfo = true
+                    })
+                } else if enableCloseAfterSeconds {
                     SBBNotification(isPresented: $isPresented3, statusType: selectedStatus, title: Text("Title"), text: Text("Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore."), maxNumberLines: enableMaxNumberLines ? maxNumberLines : nil, closeAfterSeconds: closeAfterSeconds, onMoreInfo: {
                         self.showMoreInfo = true
                     })
                 } else {
-                    SBBNotification(isPresented: $isPresented3, statusType: selectedStatus, title: Text("Title"), text: Text("Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore."), maxNumberLines: enableMaxNumberLines ? maxNumberLines : nil, closeAfterSeconds: nil, onMoreInfo: {
+                    SBBNotification(isPresented: $isPresented3, statusType: selectedStatus, title: Text("Title"), text: Text("Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore."), maxNumberLines: enableMaxNumberLines ? maxNumberLines : nil, onMoreInfo: {
                         self.showMoreInfo = true
                     })
                 }
-                if !isPresented3 {
+                if !isPresented3 && canBeClosed {
                     HStack {
                         Spacer()
                         Button(action: {
@@ -105,10 +120,24 @@ struct NotificationDemo: View {
                         Spacer()
                     }
                 }
-                                    
+                
+                if !canBeClosed {
+                    Text("With title and retry")
+                    SBBNotification(statusType: selectedStatus, title: Text("Title"), text: Text("Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore."), maxNumberLines: enableMaxNumberLines ? maxNumberLines : nil, onRetry: {
+                        self.retry = true
+                    })
+                    
+                    Text("Without title and retry")
+                    SBBNotification(statusType: selectedStatus, text: Text("Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore."), maxNumberLines: enableMaxNumberLines ? maxNumberLines : nil, onRetry: {
+                        self.retry = true
+                    })
+                }
             }
             .sbbScreenPadding()
         }
+        .alert(isPresented: $retry, content: {
+            Alert(title: Text("Retry"), message: Text("Retrying something..."), dismissButton: .cancel(Text("OK")))
+        })
         .sbbModal(isPresented: $showMoreInfo) {
             SBBModalView(title: Text("More info"), style: .bottom, isPresented: $showMoreInfo) {
                 VStack {